pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = ["Тимур Шаймарданов t.shaymardanov@gmail.com"]

[tool.poetry.dependencies]
python = "3.10.5"
Flask = "^2.2.0"
jsonschema = "^4.9.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
